+++ lib/gem_checksums/version_gem.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "version_gem"
+require_relative "version"
+
+GemChecksums::Version.class_eval do
+  extend VersionGem::Basic
+end