+++ lib/stone_checksums.rb
@@ -7,8 +7,6 @@
 #
 # RubyGems does not allow publishing a gem named `gem_checksums`,
 # hence the updated namespace and gem name.
-require "version_gem"
-
 require_relative "gem_checksums"
 
 # This library's version
@@ -50,9 +48,4 @@ module StoneChecksums
       ::GemChecksums.generate(git_dry_run: git_dry_run)
     end
   end
-end
-
-# Extend the Version module with helpers from version_gem
-StoneChecksums::Version.class_eval do
-  extend VersionGem::Basic
 end